credstore: return only unique accounts from cred stores
If may be possible for multiple credentials for the same user account to
exist in any credential store. To avoid issues with redundant account
selection prompts we should ensure duplicate accounts are removed from
the results of `ICredentialStore.GetAccounts`.

Call `.Distinct()` or use a `HashSet` where appropriate on all
implementation of `ICredentialStore.GetAccounts`.