msauth: abstract token cache init helpers
Refactor the token cache helper methods to allow us to re-use the
existing cache registration logic with a different ITokenCache and
StorageCreationProperties.

This will be useful when we later introduce a confidential client
application (for service principals) that needs a different cache
location, and uses the AppTokenCache, rather than the User one.