github: support gist remote URLs for GitHub
We have not been consistently detecting or normalising "gist" URLs for
dotcom or GHES instances.

Gists are backed by a Git repository and can be cloned/pushed-to etc
like a normal repository. Credentials are the same as the base site.

Update our OAuth, rest API, and dotcom-detection methods that deal with
the remote or target URL to correctly support gists URLs. Also add some
tests around this.