build: enable single-file publishing compression
Enable compression of assemblies in the single-file distribution of GCM.
This helps reduce the overall size on disk of GCM, which is important on
Windows as we are bundled with Git for Windows, and we don't want to
bloat their distribution.

We only enable this compression on Windows. Mac and Linux remain
uncompressed.