azrepos: do not query local bindings if outside repo
Do not attempt to query for an organisation/user binding in the local
repository configuration if we're not currently inside of a repository!

This will prevent the warning message from Git's standard error stream
from flowing to our caller's stderr stream when GCM tries to use the
`--local` option in a `git config` call.