install-from-source.sh: fix libssl Alpine pkgs
The version of libssl available in the Alpine package feeds changed with
version 3.15. The latest libssl package is 1.1 in 3.14.x and earlier,
but newer versions only have libssl3.

Add a check for the Alpine version in the install from source script and
switch between installing libssl3 and libssl1.1 depending on the distro
version.

Also add another entry to the test matrix of distributions to include a
3.14.x Alpine version.