azrepos: fix error message to stderr when not in repo
Suppress the standard error stream from Git when checking if we're
inside a repository or not, because the failure message is otherwise
printed to the user's console when we are not (which is what we're
trying to decide!).

Note that we normally don't suppress or redirect the standard error
stream for Git commands as the user may have enabled Git tracing (v1 or
v2) to the stderr file and we don't want to mute these for all calls.