Fix build in Visual Studio
Building in Visual Studio causes some tasks to be run concurrently that
would be run in a specific order in MSBuild instead. The symptom would
look like this:

	MSB4044: The "GenerateWindowsAppManifest" task was not given a
	value for the required parameter "Version".

Let's help Visual Studio realize that there are certain dependencies
between the `GetVersion` and the `GenerateWindowsAppManifest` task.

Reported by Michael J. Lyons.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>