Replace broken `PlatformFact()` constructs
We just upgraded XUnit to a newer version, which unfortunately no longer
works with the `PlatformFact()` constructs of `Xunit.SkippableFact` even
though we updated to the latest version, v1.4.13. It might have
something to do with the fact that that package has not been updated
since July 9th, 2024.

Happily, XUnit has grown equivalent features in the meantime that we can
use instead. So let's use those XUnit-native constructs instead.

Note that we still cannot drop the `SkippableFact` dependency
altogether because we need it in the `MacOSKeychain_ReadWriteDelete`
test case. It is needed to work around a flaky test that is caused by
semi-random broken states of macOS' key-chain, and that can only be
detected while the test case is running (and hence _needs_
`AssertEx.Skip()`, which in turn requires `Xunit.Skip.If()` that is
provided only via the `SkippableFact` package and there is no equivalent
native XUnit functionality).

Helped-by: Matthew Cheetham <mattche@microsoft.com>
Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>