generic: support GCM_ALLOW_UNSAFE_REMOTES option
Note that we only emit a warning for the generic host provider rather
than failing-fast like the other providers do. This is because we
never blocked HTTP remotes previously in the generic provider (which is
often used for localhost, custom hosts, etc) and don't want to break
any existing scenarios or scripts.

The new option can be used to dismiss this warning message.