credstore: add no-op credential storage option
Add a null/no-op credential store option that, as the name suggests,
does nothing. This can be useful if the user wants to use another
credential helper, configured in-front of GCM via Git, to store
credentials.

Example config:

```ini
[credential]
	credentialStore = none
	helper = /bin/my-awesome-helper
	helper = /usr/local/bin/git-credential-manager
```

In this example, the `my-awesome-helper` will be consulted first to
retrieve existing credentials before GCM, and will be asked to store
any credentials generated by GCM.