streams: only consider LF and CRLF as newlines
Git only considers LF (`\n`) and CRLF (`\r\n`) as valid line endings for
the Git credential protocol. Lone carriage-returns should be treated as
part of the value, and not a delimiter in the credential protocol.

Override the behaviour of the standard `StreamReader`'s `ReadLineAsync`
method to only break on LF or CRLF, in alignment with Git. Note that
we also override the non-async `ReadLine` method too as this is also
implemented separatley in the base class from the async version.
We must also make allowances for .NET Framework where the override
of `ReadLineAsync` that takes a `CancellationToken` does not exist.