trace2: fix pipe/socket name parsing
There are a few bugs in the way we read a pipe or socket name for the
TRACE2 event target. We do not correctly trim the \\.\pipe\ prefix on
Windows, if present. On Unix we do not correctly strip the af_unix:
prefix correctly if a socket file path contains a ':'.

Let's fix this!

Signed-off-by: Matthew John Cheetham <mjcheetham@outlook.com>