linux/pack.sh: detect runtime from arch if not specified
Detect the runtime (linux-x64/linux-arm64/etc) from the current host
architecture if not specified via the `--runtime` argument.

We had already been doing this in the script, but after we already
`die`-d when `--runtime` was missing! Oops.

Let's move this auto-detection logic up-front, and also no longer rely
on `dpkg-architecture` to determine the host arch; we can use `uname -m`
instead.

Signed-off-by: Matthew John Cheetham <mjcheetham@outlook.com>